pyproject.toml
[project]
name = "mcp-service"
version = "0.1.0"
description = "Mail Communication Platform Microservice"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.0.0",
    "httpx>=0.27.0",
    "aiosmtplib>=3.0.0",
    "google-api-python-client>=2.147.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "tenacity>=9.0.0",
    "redis>=5.0.0",
    "boto3>=1.35.0",
    "gunicorn>=21.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
    "ruff>=0.5.0"
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"